Add refactored ESMMetadataStoreCatalog
