+++ setup.cfg
@@ -10,7 +10,7 @@ select = B,C,E,F,W,T4,B9
 
 [isort]
 known_first_party=intake_esm
-known_third_party=intake,intake_xarray,numpy,pandas,pytest,setuptools,tqdm,xarray,yaml
+known_third_party=intake,intake_xarray,numpy,pandas,pytest,setuptools,xarray,yaml
 multi_line_output=3
 include_trailing_comma=True
 force_grid_wrap=0
