+++ tests/test_collections/cesm/cesm_dple_test_collection.csv
@@ -1,0 +1,7 @@
+,resource,resource_type,direct_access,experiment,case,component,stream,variable,date_range,ensemble,files,files_basename,files_dirname,ctrl_branch_year,year_offset,sequence_order,has_ocean_bgc,grid
+0,SAMPLE-DATA:posix:/Users/abanihi/devel/ncar/intake-cesm/tests/sample_data,posix,True,g.e11_LENS.GECOIAF.T62_g16.009-copy,g.e11_LENS.GECOIAF.T62_g16.009,ocn,pop.h,ECOSYS_XKW,024901-031612,0,/Users/abanihi/devel/ncar/intake-cesm/tests/sample_data/g.e11_LENS.GECOIAF.T62_g16.009.pop.h.ECOSYS_XKW.024901-031612.nc,g.e11_LENS.GECOIAF.T62_g16.009.pop.h.ECOSYS_XKW.024901-031612.nc,/Users/abanihi/devel/ncar/intake-cesm/tests/sample_data/,,1699,0,False,POP_gx1v6
+1,SAMPLE-DATA:posix:/Users/abanihi/devel/ncar/intake-cesm/tests/sample_data,posix,True,g.e11_LENS.GECOIAF.T62_g16.009-copy,g.e11_LENS.GECOIAF.T62_g16.009,ocn,pop.h.sigma,O2,024901-031612,0,/Users/abanihi/devel/ncar/intake-cesm/tests/sample_data/g.e11_LENS.GECOIAF.T62_g16.009.pop.h.sigma.O2.024901-031612.nc,g.e11_LENS.GECOIAF.T62_g16.009.pop.h.sigma.O2.024901-031612.nc,/Users/abanihi/devel/ncar/intake-cesm/tests/sample_data/,,1699,0,False,POP_gx1v6
+2,SAMPLE-DATA:posix:/Users/abanihi/devel/ncar/intake-cesm/tests/sample_data,posix,True,g.e11_LENS.GECOIAF.T62_g16.009-copy,g.e11_LENS.GECOIAF.T62_g16.009,ocn,pop.h.ecosys.nday1,CaCO3_form_zint,02490101-03161231,0,/Users/abanihi/devel/ncar/intake-cesm/tests/sample_data/g.e11_LENS.GECOIAF.T62_g16.009.pop.h.ecosys.nday1.CaCO3_form_zint.02490101-03161231.nc,g.e11_LENS.GECOIAF.T62_g16.009.pop.h.ecosys.nday1.CaCO3_form_zint.02490101-03161231.nc,/Users/abanihi/devel/ncar/intake-cesm/tests/sample_data/,,1699,0,False,POP_gx1v6
+3,TEST-LIST:input-file:/Users/abanihi/devel/ncar/intake-cesm/tests/intake-cesm-test-filelist,input-file,False,g.e11_LENS.GECOIAF.T62_g16.009,g.e11_LENS.GECOIAF.T62_g16.009,ocn,pop.h.ecosys.nday1,CaCO3_form_zint,02490101-03161231,0,example_input/sample_data/g.e11_LENS.GECOIAF.T62_g16.009.pop.h.ecosys.nday1.CaCO3_form_zint.02490101-03161231.nc,g.e11_LENS.GECOIAF.T62_g16.009.pop.h.ecosys.nday1.CaCO3_form_zint.02490101-03161231.nc,example_input/sample_data/,,1699,0,False,POP_gx1v6
+4,TEST-LIST:input-file:/Users/abanihi/devel/ncar/intake-cesm/tests/intake-cesm-test-filelist,input-file,False,g.e11_LENS.GECOIAF.T62_g16.009,g.e11_LENS.GECOIAF.T62_g16.009,ocn,pop.h.sigma,O2,024901-031612,0,example_input/sample_data/g.e11_LENS.GECOIAF.T62_g16.009.pop.h.sigma.O2.024901-031612.nc,g.e11_LENS.GECOIAF.T62_g16.009.pop.h.sigma.O2.024901-031612.nc,example_input/sample_data/,,1699,0,False,POP_gx1v6
+5,TEST-LIST:input-file:/Users/abanihi/devel/ncar/intake-cesm/tests/intake-cesm-test-filelist,input-file,False,g.e11_LENS.GECOIAF.T62_g16.009,g.e11_LENS.GECOIAF.T62_g16.009,ocn,pop.h,ECOSYS_XKW,024901-031612,0,example_input/sample_data/g.e11_LENS.GECOIAF.T62_g16.009.pop.h.ECOSYS_XKW.024901-031612.nc,g.e11_LENS.GECOIAF.T62_g16.009.pop.h.ECOSYS_XKW.024901-031612.nc,example_input/sample_data/,,1699,0,False,POP_gx1v6
