+++ tests/test_core.py
@@ -6,48 +6,42 @@ import pytest
 import xarray as xr
 
 # pytest imports this package last, so plugin is not auto-added
-from intake_esm.core import CesmMetadataStoreCatalog
+from intake_esm.core import ESMMetadataStoreCatalog
 
 here = os.path.abspath(os.path.dirname(__file__))
 
 
 def test_build_collection():
     collection_input_file = os.path.join(here, 'collection_input_test.yml')
-    build_args = {
-        'collection_input_file': collection_input_file,
-        'collection_type': "cesm",
-        'overwrite_existing': True,
-    }
-    col = intake.open_cesm_metadatastore(collection='test', build_args=build_args)
+    col = intake.open_esm_metadatastore(collection_input_file)
+    col = col.open_collection("cesm_dple_test_collection", collection_type="cesm")
     assert isinstance(col.df, pd.DataFrame)
 
 
-@pytest.mark.skip
 @pytest.mark.parametrize("collection", ["cesm_dple", "cesm_dple_test_collection"])
 def test_constructor(collection):
-    c = intake.open_cesm_metadatastore(collection)
-    assert isinstance(c, CesmMetadataStoreCatalog)
+    c = intake.open_esm_metadatastore(collection_name=collection, collection_type="cesm")
+    assert isinstance(c, ESMMetadataStoreCatalog)
 
 
-@pytest.mark.skip
 def test_search():
-    c = intake.open_cesm_metadatastore("cesm_dple")
+    c = intake.open_esm_metadatastore(collection_name="cesm_dple", collection_type="cesm")
     cat = c.search(experiment="g.e11_LENS.GECOIAF.T62_g16.009", component="ocn", variable="FG_CO2")
 
     assert isinstance(cat.results, pd.DataFrame)
     assert not cat.results.empty
 
 
-@pytest.mark.skip
 def test_cat():
     cat = intake.open_catalog(os.path.join(here, "catalog.yaml"))
-    cat = cat["cesm_dple-cff53aef-6938-4c6e-b6ae-efa5035bed7e"]
+    cat = cat["cesm_dple_test_collection_7afe8a3a-8d2f-40a8-9ede-899f48ce83b2"]
     assert isinstance(cat.results, pd.DataFrame)
 
 
-@pytest.mark.skip
 def test_to_xarray():
-    c = intake.open_cesm_metadatastore("cesm_dple_test_collection")
+    c = intake.open_esm_metadatastore(
+        collection_name="cesm_dple_test_collection", collection_type="cesm"
+    )
     cat = c.search(variable='O2', direct_access=True)
     ds = cat.to_xarray()
     assert isinstance(ds, xr.Dataset)
